refactor(L2): rpc module (#731)
- Rename `l2/rpc` module to `l2/eth_client`.
- Move `eth_client` module to `l2/utils`.
- Rename `L1Rpc` -> `EthClient`.
