+++ crates/l2/l2.rs
@@ -1,6 +1,6 @@
 pub mod operator;
 pub mod prover;
-pub mod rpc;
+pub mod utils;
 
 pub use operator::start_operator;
 pub use prover::start_prover;
