--- crates/l2/rpc/mod.rs
@@ -1,2 +1,0 @@
-pub mod l1_rpc;
-pub mod transaction;
