+++ crates/l2/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod eth_client;
